# [libc++] Speed up classic locale (take 2)