@ldionne 2023-11-28
Re-purposing this PR to re-apply #72112 which was reverted.